+++ packages/generator/lib/combo/patch-build/index.ts
@@ -72,7 +72,6 @@ function asmPatchGroups(world: LogicResultWorld, settings: Settings) {
     CRIT_WIGGLE_DISABLE: settings.critWiggleDisable,
     OOT_LOST_WOODS_EXITS: settings.alterLostWoodsExits,
     MM_VOID_WARP: settings.voidWarpMm,
-    OOT_SWORDLESS_ADULT: settings.swordlessAdult,
     MM_FAST_MASKS: settings.fastMasks,
     OOT_FAST_BOAT: settings.shadowFastBoat,
     MM_WELL_OPEN_OR_REMORSELESS: settings.beneathWell !== 'vanilla',